+++ container_server/src/cmmp/InputTruckResponsePacket.cpp
@@ -12,14 +12,14 @@ InputTruckResponsePacket InputTruckResponsePacket::decode(std::vector<char>::con
         if (size) {
             std::string container_id;
             std::string destination;
-            uint32_t x;
-            uint32_t y;
+            uint16_t x;
+            uint16_t y;
 
             for(uint32_t i = 0; i < size; i++) {
                 container_id = readString(it);
                 destination = readString(it);
-                x = readPrimitive<uint32_t>(it);
-                y = readPrimitive<uint32_t>(it);
+                x = readPrimitive<uint16_t>(it);
+                y = readPrimitive<uint16_t>(it);
 
                 containers.emplace_back(Container { container_id, destination, x, y });
             }
